fix: bug in updating task duration.
Issue: The task duration being updated on the basis of task name which causes to update the duration of older days too.
Solution: The bug is fixed by updating the task duration on the basis of both task name and current date which will only causes the update of task duration of current day not all days.